card names healpers function
